requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
wtforms==3.0.1
email-validator==2.0.0
flask-migrate==4.0.5
python-dotenv==1.0.0
werkzeug==2.3.7
flask-bcrypt==1.0.1
Pillow==10.0.0
watchdog==2.3.1